Backend/requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
django-model-utils==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
psycopg2-binary==2.9.3